:green_heart: Merge IT test
Until spec can be run sequentially
