+++ src/test/scala/zio/pravega/test/PravegaContainer.scala
@@ -29,7 +29,7 @@ class PravegaContainer(
 object PravegaContainer {
   def pravega: ZLayer[Scope, Nothing, PravegaContainer] = {
     val imageName = sys.env.getOrElse("PRAVEGA_IMAGE", "pravega/pravega:0.10.2")
-    ZIO.acquireRelease {
+    ZLayer(ZIO.acquireRelease {
       ZIO.attemptBlocking {
         val container = new PravegaContainer(
           dockerImageName = DockerImageName.parse(imageName)
